+++ src/main/java/com/project/webshopproject/common/RestApiResponseDto.java
@@ -1,4 +1,4 @@
-package com.project.webshopproject.global;
+package com.project.webshopproject.common;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
